requirements.txt
appnope==0.1.4
asttokens==3.0.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.58.0
geoprop @ file:///Users/patrick/Dev/helium-iot-classifiers/geoprop-py/target/wheels/geoprop-0.1.0-cp313-cp313-macosx_11_0_arm64.whl#sha256=54315a12a693010e8cfa75ebc8858ccd1cbce0f23a7c9d0b817a88724c263f1d
h3==4.2.2
idna==3.10
importlib==1.0.4
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
polars==1.29.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-arango==8.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==26.4.0
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.4.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
urllib3==2.4.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
zipp==3.21.0
